agrego cálculo de tiempos en burbujeo

+++ src/src/Paso.java
@@ -9,7 +9,7 @@ public class Paso {
 	private double tiempoTranscurrido;
 	
 	
-	public Paso(TipoPaso tipo, int tiempoTranscurrido, int posElem1, int posElem2) {
+	public Paso(TipoPaso tipo, long tiempoTranscurrido, int posElem1, int posElem2) {
 		super();
 		this.tipo = tipo;			
 		this.tiempoTranscurrido = tiempoTranscurrido;
